requirements.txt
flask>=2.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pdfminer.six>=20221105
python-docx>=0.8.11
reportlab>=4.0.0

